requirements.txt
slack_bolt
flask
python-dotenv
gspread
google-auth